[Refactor] Package refactoring / binding grouping
