+++ app/src/main/res/layout/fragment_recipe.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="reciepviewmodel"
-            type="com.angdroid.refrigerator_manament.presentation.home.RecipeViewModel" />
+            type="com.angdroid.refrigerator_manament.presentation.home.viewmodel.RecipeViewModel" />
 
     </data>
